move BackupFileListBuilder from test to testtools
test/backup_file_list_builder_for_test.go -> testtools/backup_file_list_builder.go
